[libc] Split 'malloc'
