+++ libc/malloc/__alloca_alloc.c
@@ -1,0 +1,3 @@
+#include <malloc.h>
+
+void *(*__alloca_alloc) __P ((size_t)) = __mini_malloc;
