+++ libc/malloc/__freed_list.c
@@ -1,0 +1,3 @@
+#include "_malloc.h"
+
+mem  *__freed_list = 0;
