+++ libc/malloc/brk.c
@@ -1,0 +1,8 @@
+#include <unistd.h>
+
+int _brk (void *);
+
+int brk (void * addr)
+{
+   return _brk (addr);
+}
